+++ NEWS.md
@@ -1,9 +1,22 @@
-# DEbChIP 0.2.0
+# DEbPeak 0.4.0
 ## New features
-* Add `FindMotif` to find motif in interested peaks (suitable for RNA-seq and ATAC-seq integration).
+* Added `DEbCA` to support integrating RNA-seq, ChIP-seq and ATAC-seq.
+* Modified `PlotDEbPeak` to support integrating RNA-seq, ChIP-seq and ATAC-seq.
 
 -------------
 
-# DEbChIP 0.1.0
+# DEbPeak 0.3.0
+## New features
+* Changed package name to `DEbPeak` to broaden its usage for ChIP-seq, ATAC-seq.
+
+-------------
+
+# DEbPeak 0.2.0
+## New features
+* Added `FindMotif` to find motif in interested peaks (suitable for RNA-seq and ATAC-seq integration).
+
+-------------
+
+# DEbPeak 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.